+++ Sources/Plasma/FeatureLib/pfGameGUIMgr/pfGUIProgressCtrl.cpp
@@ -62,7 +62,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 
 #include "pnSceneObject/plSceneObject.h"
 #include "pnSceneObject/plCoordinateInterface.h"
-#include "pnTimer/plTimerCallbackManager.h"
+#include "plTimerCallbackManager.h"
 
 #include "plgDispatch.h"
 #include "hsResMgr.h"
