+++ Sources/Plasma/NucleusLib/inc/CMakeLists.txt
@@ -23,8 +23,10 @@ set(pnNucleusInc_HEADERS
     plProfile.h
     plProfileManager.h
     plRefFlags.h
+	plTimerCallbackManager.h
     pnAllCreatables.h
     pnNucleusCreatables.h
+	pnTimerCreatable.h
 )
 
 set(pnNucleusInc_SOURCES
@@ -33,7 +35,7 @@ set(pnNucleusInc_SOURCES
 )
 
 add_library(pnNucleusInc STATIC ${pnNucleusInc_HEADERS} ${pnNucleusInc_SOURCES})
-target_link_libraries(pnNucleusInc CoreLib pnTimer)
+target_link_libraries(pnNucleusInc CoreLib)
 
 source_group("Header Files" FILES ${pnNucleusInc_HEADERS})
 source_group("Source Files" FILES ${pnNucleusInc_SOURCES})
